+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.0.2 - 5th November 2024
+## 0.0.2
 
 ### Added
 
@@ -6,3 +6,44 @@
 - Implemented manual control for speech-to-text functionality with start, stop, pause, and resume capabilities.
 - Added callbacks for pause and play actions.
 - Included an example Flutter app demonstrating package usage.
+
+## 0.0.3
+
+### Enhancements
+
+- Added CHANGELOG.md.
+- Enhancement in documentation.
+
+## 0.0.4 & 0.0.5 & 0.0.6
+
+### Enhancements
+
+- Package description enhancement.
+
+## 0.0.7 & 0.0.8
+
+### Enhancements
+
+- While speaking, real time text state will emit now.
+- Enhancements in keywords
+- enhancements in README.md
+
+## 0.0.9
+
+### Enhancements
+
+- Added manifest file code in android module
+- Added permanently denied microphone permission handeling (added default dialog, and overrided callback for same).
+
+## 0.0.10 && 0.0.11 && 0.0.12 && 0.0.13 && 0.0.14
+
+### Enhancements
+
+- Updated `AndroidManifest.xml` path.
+- Trying to add a manifest file and its code from package side only but didn't work.
+
+## 0.0.15
+
+### Enhancements
+
+- Now removed `AndroidManifest.xml`, so now to use this package, please add necessary code in `AndroidManifest.xml`.